[Update] Preliminary Login Update
git-svn-id: http://svn2.assembla.com/svn/swgemu/core3/branches/Object%20Revision@1583 c3d1530f-68f5-4bd0-87dc-8ef779617e40

+++ src/server/zone/objects/tangible/TangibleObjectImplementation.h
@@ -158,7 +158,7 @@ public:
 	}
 
 	inline void setConditionMax(float max) {
-		conditionMax = max;
+		conditionMax = (uint32)max;
 	}
 
 	//Getters
